Change to automatically rebuild views on error
Previously FlumeDB would handle errors by retrying the view build
operation, which would usually fail recursively until it hit the maximum
stack size. This behavior meant that "delete your views" was a common
fix, which is difficult for beginners and error-prone for everyone.
Running `rm -rf` in your database directory is scary.

This commit changes the behavior so that errors while building views
will trigger a rebuild instead of recursively retrying the build. This
gives FlumeDB the ability to self-heal when a Flumeview becomes corrupt,
but (as always) is a trade-off. Some errors are probably harmless or
will resolve themselves without manual intervention, but FlumeDB
shouldn't be responsible for enumerating the possible error states of
each Flumeview. Each Flumeview should maintain responsibility for its
own error states.

For example, if two instances of Flumeview-Level try to use the same
directory, we can expect the second instance to throw an error because
the first has locked the directory. If FlumeDB asks the second instance
to rebuild, it should throw a rebuild error and refuse to write to the
locked directory. As long as Flumeviews continue to refuse rebuilds when
manual intervention is unnecessary I think this change will be a big
improvement with a low risk of problematic behavior.