Convert to StandardJS
Problem: Having an inconsistent style makes the code harder to use,
improve, and especially diff between branches.

Solution: I like using code style tools like Prettier, but in the
subculture surrounding FlumeDB it's more common to use StandardJS, which
does both code style and linting. This commit was generated with:

    npx prettier --write . && npx standard --fix

Which converted to Prettier syntax and then flipped that into the
StandardJS style. This means that it also fixed our Markdown.