Resolve test/rebuild.js failure
Problem: There's a problem where using `/tmp/` as the parent directory
makes some files clash, which triggers a rebuild as soon as
Flumeview-Level starts. It seems to be a problem with Flumelog-Offset,
but the stream restart happens in Flumeview-Level. I don't understand
the problem completely.

Solution: Use `/tmp/test-flumedb-<random/` as the parent directory to
avoid filename clashes that fail the tests.