Add abort() to rebuild to cancel old streams
Problem: During a rebuild we restart the view stream to `createSink()`
as soon as the previous stream closes. This creates a race condition
where the view has to choose between emptying the database and closing
the stream. If you close the stream first, the stream will restart
before you finish emptying the database. If you empty the database
first, you may receive new messages before the old stream is cancelled.
Handling all of the edge-cases in each individual view is hard.

Solution: Make FlumeDB abort the view streams to `createSink()` at the
beginning of the rebuild process, and be especially careful to prevent
the stream from restarting until the view's `destroy()` method is
completely finished. Managing this in FlumeDB should make it easier to
prevent race conditions and confusing database states.